Rename all wordlist files to contextpack files
