+++ client/src/app/app-routing.module.ts
@@ -4,9 +4,9 @@ import { HomeComponent } from './home/home.component';
 import { UserListComponent } from './users/user-list.component';
 import { UserProfileComponent } from './users/user-profile.component';
 import { AddUserComponent } from './users/add-user.component';
-import { AddWordlistComponent } from './wordlists/add-wordlist.component';
-import { WordlistListComponent } from './wordlists/wordlist-list.component';
-import { WordlistInfoComponent } from './wordlists/wordlist-info.component';
+import { AddWordlistComponent } from '../wordlists/contextpack-list.component
+import { WordlistListComponent } from './contextpacks/wordlist-list.component';
+import { WordlistInfoComponent } from './contextpacks/contextpack-info.component';
