+++ client/src/app/contextpacks/contextpack-card.component.html
