+++ client/src/app/contextpacks/contextpack-info.component.spec.ts
@@ -4,10 +4,10 @@ import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ActivatedRouteStub } from '../../testing/activated-route-stub';
 import { MockWordlistService } from '../../testing/wordlist.service.mock';
-import { Wordlist } from './wordlist';
-import { WordlistCardComponent } from './wordlist-card.component';
-import { WordlistInfoComponent } from './wordlist-info.component';
-import { WordlistService } from './wordlist.service';
+import { Wordlist } from './contextpack';
+import { WordlistCardComponent } from './contextpack-card.component';
+import { WordlistInfoComponent } from './contextpack-info.component';
+import { WordlistService } from './contextpack.service';
 
 describe('WordlistInfoComponent', () => {
   let component: WordlistInfoComponent;
