+++ client/src/app/contextpacks/contextpack-info.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Wordlist } from './wordlist';
-import { WordlistService } from './wordlist.service';
+import { Wordlist } from './contextpack';
+import { WordlistService } from './contextpack.service';
 import { Subscription } from 'rxjs';
 
 @Component({
