+++ client/src/app/contextpacks/contextpack.service.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
-import { Wordlist } from './wordlist';
+import { Wordlist } from './contextpack';
 import { map } from 'rxjs/operators';
 
 @Injectable()
